Music Recommendation using Chatbot/Chatbot/requirement.txt
numpy==1.24.3
nltk==3.8.1
Keras==2.12.0
tensorflow==2.12.0
azureml-core==1.50.0
